+++ script/build.py
@@ -5,7 +5,7 @@ if __name__ == '__main__':
     opts = [
         '--onefile',
         '--windowed',
-        '--icon=images/company_logo.png',
+        '--icon=images/logo.jpg',
         'main.py'
     ]
     sys.exit(run(opts))